refactor(structure): split paymentFlowService/ArticleComposer/cron routes + size gate
Pure moves only — zero behavior change:
- paymentFlowService.ts (969) -> facade over paymentInitiation /
  paymentStatusFlow / paymentSettlement / paymentFlowHelpers
- ArticleComposer.tsx (483) -> ComposerEditor, ArticlePreview, DraftsPanel,
  PublishBar + useLocalArticleDrafts hook
- cron route bodies -> services/payments/reconcile.ts and
  services/webhooks/worker.ts; routes are now secret-check + call + respond
- scripts/check-file-sizes.js wired into verify (components>300,
  routes>150, services>500) with a shrink-only exception list seeded with
  current violators; supersedes check-long-files.js
