+++ package.json
@@ -31,7 +31,8 @@
     "lint:check": "next lint",
     "lint:umlauts": "grep -rn --include='*.ts' --include='*.tsx' -E '(fuer|koennen|moeglich|muessen|wuerde|aehnlich|aerger|aendern|oeffnen|ueber|ueberall|zurueck|natuerlich|genuegt|Strasse|Gruesse)' src/ || echo 'No umlaut violations found'",
     "type-check": "tsc --noEmit --skipLibCheck --incremental false",
-    "verify": "npm run ci:docs && npm run type-check && npm run audit:routes && npm run lint && npm run test:unit -- --watchAll=false",
+    "check:sizes": "node scripts/check-file-sizes.js",
+    "verify": "npm run ci:docs && npm run type-check && npm run check:sizes && npm run audit:routes && npm run lint && npm run test:unit -- --watchAll=false",
     "audit:schema": "node scripts/db/audit-schema-drift.mjs",
     "audit:routes": "node scripts/audit-routes.mjs",
     "gen:types": "bash scripts/db/gen-types.sh",
@@ -90,7 +91,7 @@
     "bundle:analyze": "ANALYZE=true npm run build",
     "bundle:monitor": "node scripts/bundle-size-monitor.js",
     "bundle:check": "npm run build && npm run bundle:monitor",
-    "lint:long": "node scripts/check-long-files.js",
+    "lint:long": "node scripts/check-file-sizes.js",
     "cleanup:repo": "node scripts/cleanup-repo.js",
     "test:smoke": "BASE_URL=http://localhost:3000 playwright test tests/e2e/smoke.spec.ts --reporter=list",
     "test:e2e:matrix": "playwright test tests/e2e/workflow-matrix.spec.ts --project=chromium --reporter=line",
